+++ src/main/java/umc/stockoneqback/auth/domain/model/jwt/Token.java
@@ -1,4 +1,4 @@
-package umc.stockoneqback.auth.domain.model;
+package umc.stockoneqback.auth.domain.model.jwt;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
