+++ src/test/java/umc/stockoneqback/auth/service/AuthServiceTest.java
@@ -2,8 +2,7 @@ package umc.stockoneqback.auth.service;
 
 import org.junit.jupiter.api.*;
 import org.springframework.beans.factory.annotation.Autowired;
-import umc.stockoneqback.auth.domain.model.FcmToken;
-import umc.stockoneqback.auth.domain.model.RefreshToken;
+import umc.stockoneqback.auth.domain.model.fcm.FcmToken;
 import umc.stockoneqback.auth.exception.AuthErrorCode;
 import umc.stockoneqback.auth.service.dto.response.LoginResponse;
 import umc.stockoneqback.auth.service.jwt.AuthService;
